Roll back some name changes; code cleanup for more consistency
