+++ Libplanet.Explorer/GraphTypes/BoundPeerType.cs
@@ -7,6 +7,8 @@ namespace Libplanet.Explorer.GraphTypes
     {
         public BoundPeerType()
         {
+            Name = "BoundPeer";
+
             // We need multiple row of description for clearer, not confusing explanation of field.
             Field<NonNullGraphType<PublicKeyType>>(
                 "publicKey",
@@ -23,8 +25,6 @@ namespace Libplanet.Explorer.GraphTypes
                 description: "The address of the miner.",
                 resolve: ctx => ctx.Source.PublicIPAddress?.ToString()
             );
-
-            Name = "BoundPeer";
         }
     }
 }
